Fix broken compilation in linux
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3280 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/extra/word_generator/main_word_generator.cpp
@@ -124,7 +124,7 @@ int main( int argC , const char*argV[] )
     {
         // Check the limit of generated words
         if( max_num_lines > 0 )
-            if( num_lines >= max_num_lines )
+		   if( num_lines >= (size_t) max_num_lines )
                 break; 
         
         // Get new word
